# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GenXcare API
  version: 1.0.0
  description: REST contract for the GenXcare modular monolith.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { $ref: "#/components/responses/Page" }, default: { $ref: "#/components/responses/Error" } }
  /v1/notifications/preferences:
    get:
      tags: [Notifications]
      operationId: getNotificationPreferences
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
    patch:
      tags: [Notifications]
      operationId: updateNotificationPreferences
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/notifications/devices:
    post:
      tags: [Notifications]
      operationId: registerNotificationDevice
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
...
  /v1/notifications/reminders:
    get:
      tags: [Notifications]
      operationId: getReminderPlan
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllNotificationsRead
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
...
  requestBodies:
    JsonObject:
      required: true
      content: { application/json: { schema: { type: object } } }
...
  responses:
...
    Entity:
      description: Resource
      content: { application/json: { schema: { type: object } } }
...
    Page:
      description: Cursor page
      content: { application/json: { schema: { $ref: "#/components/schemas/PageEnvelope" } } }
...
    Error:
      description: Request failed
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
...
    ListEnvelope:
      type: object
      required: [items]
      properties: { items: { type: array, items: { type: object } } }
    PageEnvelope:
      allOf:
        - { $ref: "#/components/schemas/ListEnvelope" }
        - { type: object, required: [nextCursor], properties: { nextCursor: { type: [string, "null"] } } }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: { type: string } }